refactor: remove old code and replace with octree code
