--- src/new_world/mod.rs
@@ -1,5 +1,0 @@
-pub mod block;
-pub mod chunk;
-mod generate;
-mod octree;
-pub mod world;
