+++ src/util/mod.rs
@@ -1,1 +1,2 @@
+pub mod octree;
 pub mod primitives;
