--- src/world/ecs/mod.rs
@@ -1,3 +1,0 @@
-pub mod chunk_loader;
-pub mod physics;
-pub mod player;
